refactor: Inline sortByRank
